requirements.txt
streamlit
numpy
pandas
plotly_express
plotnine
scipy.stats